server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
transformers
pymilvus
google-genai
python-dotenv
pydantic
supabase
requests
